Lesson_4_1. Implementation of prop types
