+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "airtable": "^0.12.2",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.21.2",
