+++ src/components/TodoList.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import TodoListItem from './TodoListItem';
 import styles from './TodoList.module.css';
 
@@ -20,6 +21,16 @@ const TodoList = ({ todoList, onRemoveTodo, onToggleTodoCompletion }) => {
         </>
     );
 }
+
+TodoList.propTypes = {
+    todoList: PropTypes.arrayOf(PropTypes.shape({
+        id: PropTypes.string.isRequired,
+        title: PropTypes.string.isRequired,
+        isCompleted: PropTypes.bool
+    })).isRequired,
+    onRemoveTodo: PropTypes.func.isRequired,
+    onToggleTodoCompletion: PropTypes.func.isRequired
+};
 
 export default TodoList;
